[Refactor] Extract workflow/workflow tabs restore into a composable
